requirements.txt
streamlit==1.51.0
google-generativeai==0.3.0
opencv-python==4.12.0.0
Pillow==12.0.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.24.3
